requirements.txt
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
yfinance>=0.2.0
hydra-core>=1.3.0
omegaconf>=2.3.0
pyarrow>=12.0.0

